+++ Forms/CommandLineForm.vb
@@ -1,4 +1,5 @@
-﻿Imports System.Threading
+﻿
+Imports System.Threading
 Imports System.Threading.Tasks
 
 Imports StaxRip.CommandLine