+++ Forms/MainForm.vb
@@ -1381,26 +1381,14 @@ Public Class MainForm
         PopulateProfileMenu(DynamicMenuItemID.Audio2Profiles)
         PopulateProfileMenu(DynamicMenuItemID.FilterSetupProfiles)
 
-        For Each i In CustomMainMenu.MenuItems
-            If i.CustomMenuItem.MethodName = "DynamicMenuItem" Then
-                If i.CustomMenuItem.Parameters(0).Equals(DynamicMenuItemID.HelpApplications) Then
-                    i.DropDownItems.ClearAndDisplose
+        For Each iMenuItem In CustomMainMenu.MenuItems
+            If iMenuItem.CustomMenuItem.MethodName = "DynamicMenuItem" Then
+                If iMenuItem.CustomMenuItem.Parameters(0).Equals(DynamicMenuItemID.HelpApplications) Then
+                    iMenuItem.DropDownItems.ClearAndDisplose
 
-                    For Each iPackage In Package.Items.Values
-                        If iPackage.GetHelpPath <> "" Then
-                            Dim plugin = TryCast(iPackage, PluginPackage)
-
-                            If plugin Is Nothing Then
-                                ActionMenuItem.Add(i.DropDownItems, iPackage.Name.Substring(0, 1).Upper + " | " + iPackage.Name, Sub() iPackage.ShowHelp())
-                            Else
-                                If plugin.AvsFilterNames?.Length > 0 Then
-                                    ActionMenuItem.Add(i.DropDownItems, iPackage.Name.Substring(0, 1).Upper + " | " + iPackage.Name + " (AviSynth)", Sub() g.StartProcess(iPackage.GetHelpPath(ScriptEngine.AviSynth)))
-                                End If
-
-                                If plugin.VSFilterNames?.Length > 0 Then
-                                    ActionMenuItem.Add(i.DropDownItems, iPackage.Name.Substring(0, 1).Upper + " | " + iPackage.Name + " (VapourSynth)", Sub() g.StartProcess(iPackage.GetHelpPath(ScriptEngine.VapourSynth)))
-                                End If
-                            End If
+                    For Each pack In Package.Items.Values
+                        If pack.GetHelpPath <> "" Then
+                            ActionMenuItem.Add(iMenuItem.DropDownItems, pack.Name.Substring(0, 1).Upper + " | " + pack.Name, Sub() pack.ShowHelp())
                         End If
                     Next
                 End If